infra_base/requirements.txt
boto3==1.34.11
PyYAML==5.4.1
python-dotenv==1.0.0